LPS-178068 Fix broken package naming
