+++ workspaces/liferay-partner-workspace/client-extensions/salesforce-sync-api/src/main/java/com/liferay/partner/portal/salesforce/sync/api/config/ClientIdValidator.java
